move protocol checking function next to the protocol
